Minor improvements

+++ natural-language-understanding/src/main/java/com/ibm/watson/developer_cloud/natural_language_understanding/v1/model/CategoriesResult.java
@@ -12,7 +12,6 @@
  */
 package com.ibm.watson.developer_cloud.natural_language_understanding.v1.model;
 
-import java.math.BigDecimal;
 
 import com.ibm.watson.developer_cloud.natural_language_understanding.v1.NaturalLanguageUnderstanding;
 import com.ibm.watson.developer_cloud.service.model.GenericModel;
@@ -25,7 +24,7 @@ public class CategoriesResult extends GenericModel {
   /** The path to the category through the taxonomy hierarchy. */
   private String label;
   /** Confidence score for the category classification. Higher values indicate greater confidence. */
-  private BigDecimal score;
+  private Long score;
 
   /**
    * Instantiates a new `CategoriesResult`
@@ -33,7 +32,7 @@ public class CategoriesResult extends GenericModel {
    * @param label The path to the category through the taxonomy hierarchy.
    * @param score Confidence score for the category classification. Higher values indicate greater confidence.
    */
-  public CategoriesResult(final String label, final BigDecimal score) {
+  public CategoriesResult(final String label, final Long score) {
     this.label = label;
     this.score = score;
   }
@@ -52,7 +51,7 @@ public class CategoriesResult extends GenericModel {
    *
    * @return the score
    */
-  public BigDecimal getScore() {
+  public Long getScore() {
     return score;
   }
 
@@ -70,7 +69,7 @@ public class CategoriesResult extends GenericModel {
    *
    * @param score the new score
    */
-  public void setScore(final BigDecimal score) {
+  public void setScore(final Long score) {
     this.score = score;
   }
